+++ issue-splitter-operator/src/js/main.js
@@ -13,7 +13,7 @@
     var data = null;
 
     //Initialize as empty data.
-    var statusList = [], priorityList = [], assigneeList = [], typeList = [];
+    var statusList = [], priorityList = [], assigneeList = [], typeList = [], sprintList = [], monthList = [];
 
     var init = function init() {
         MashupPlatform.wiring.registerCallback("issue-list", function (values) {
@@ -29,25 +29,32 @@
         priorityList = [];
         assigneeList = [];
         typeList = [];
+        sprintList = [];
+        monthList = [];
 
         //Splits the data
         data.forEach(function (issue) {
             if (issue.jira && issue.jira.priority) {
                 priorityList.push(issue.jira.priority);
             }
-
             statusList.push(issue.status);
             assigneeList.push(issue.assignee);
             typeList.push(issue.type);
+            sprintList.push(issue.versions);
+            monthList.push(issue.month);
         });
+
         pushInfo();
     };
 
+    //Pushes current data
     var pushInfo = function pushInfo() {
         MashupPlatform.wiring.pushEvent("status-list", statusList);
         MashupPlatform.wiring.pushEvent("priority-list", priorityList);
         MashupPlatform.wiring.pushEvent("assignee-list", assigneeList);
         MashupPlatform.wiring.pushEvent("type-list", typeList);
+        MashupPlatform.wiring.pushEvent("sprint-list", sprintList);
+        MashupPlatform.wiring.pushEvent("month-list", monthList);
     };
 
     init();